Fixed more package rename stuff
