+++ app/src/main/java/com/freshollie/monkeyboard/keystoneradio/radio/RadioDevice.java
@@ -6,7 +6,7 @@ import android.os.Handler;
 import android.os.SystemClock;
 import android.util.Log;
 
-import com.freshollie.monkeyboard.R;
+import com.freshollie.monkeyboard.keystoneradio.R;
 
 import java.nio.ByteBuffer;
 import java.nio.charset.Charset;
